refactor: break a huge file down
